pyproject.toml
[project]
name = "custom-sso-provider"
version = "0.1.0"
description = "Google-shape OIDC provider on custom issuer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.2",
  "psycopg[binary]>=3.2.1",
  "authlib>=1.3.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt==4.0.1",
  "python-multipart>=0.0.9",
  "itsdangerous>=2.2.0",
  "jinja2>=3.1.4",
  "pydantic-settings>=2.3.4",
  "typer>=0.12.3",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.2",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
